Prevent rate limiting by StatusCake (#5)
To prevent rate limiting with StatusCake two new things have
been introduced. First of all, we cache the results of the
GetAll function, so that it does not have to retrieve the status
monitors everytime.

Secondly, I added a latency after adding a monitor. Because even
when the results are cached, if several monitors have to be added
in a row, it will be invalidated and still hit the rate limit.