Statuscake: Implement Equals function using TestTags (#610)
* Implement Equals function using TestTags

It is mentioned in a comment in the code itself but because of the
discrepency between the fields in the EndpointMonitor CR and the
Statuscake API it is not immediately clear how to compare an old monitor
with an updated monitor. The way I have elected to check this is to use
the TestTags field to include some kind of identifier that should be
updated on any change. So if the tags have changed then the monitor
should be updated.

* Remove already merged content